+++ app/Http/Controllers/Api/Company/Auth/CompanyLoginController.php
@@ -1,26 +1,33 @@
 <?php
 
-namespace App\Http\Controllers\api\researcher;
+namespace App\Http\Controllers\Api\Company\Auth;
 
 use App\Http\Controllers\Controller;
 use App\Http\Traits\GeneralTrait;
 use Illuminate\Support\Facades\Auth;
-use  App\Models\Researcher;
+use App\Models\Company;
 use Illuminate\Support\Facades\Validator;
-use App\Http\Resources\ResearcherResource;
+use App\Http\Resources\CompanyResource\CompanyResource as CompanyResourceCompanyResource;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
 
-class ResearcherLoginController extends Controller
+class CompanyLoginController extends Controller
 {
+
     use GeneralTrait;
 
     /**
      * Display a listing of the resource.
      */
+
     public function index()
     {
         //
     }
+
+    /**
+     * Store a newly created resource in storage.
+     */
     public function login(Request $request)
     {
         $rules = [
@@ -28,7 +35,7 @@ class ResearcherLoginController extends Controller
                 'required',
                 'string',
                 'email',
-                'exists:researchers,email',
+                'exists:companies,email',  
             ],
             'password' => [
                 'required',
@@ -51,73 +58,50 @@ class ResearcherLoginController extends Controller
 
         $validator = Validator::make($request->all(), $rules, $messages);
 
-        if ($validator->fails()) {
+        if ($validator->fails()) 
+        {
             $firstError = $validator->errors()->first();
-
-            if (strpos($firstError, 'مطلوب') !== false) {
-                return $this->requiredField($firstError);
+        
+            if (strpos($firstError, 'مطلوب') !== false) 
+            {
+                return $this->requiredField($firstError);  
             }
 
             return $this->notFoundResponse($firstError);
         }
 
-        $researcher = Researcher::where('email', $request->email)->first();
+        $company = Company::where('email', $request->email)->first();
 
-        if (!$researcher || !\Hash::check($request->password, $researcher->password)) {
+        if (!$company || !Hash::check($request->password, $company->password))
+        {
             return $this->unAuthorizeResponse(); // بيانات الاعتماد غير صحيحة
         }
 
-        if (is_null($researcher->code)) {
-            return $this->notFoundResponse('الحساب غير مفعل عليك إدخال الكود لتفعيله');
-        }
-
-        // if ($researcher->tokens()->exists()) {
+    
+        
+        // if ($company->tokens()->exists()) 
+        // {
         //     return $this->unAuthorizeResponse();
         // }
 
-        $token = $researcher->createToken('auth_token')->plainTextToken;
+        $token = $company->createToken('auth_token')->plainTextToken;
 
-        return (new ResearcherResource($researcher))->successResponseWithToken($token);
+        return (new CompanyResourceCompanyResource($company))->successResponseWithToken($token);
     }
-
+    
     public function logout()
     {
-        $researcher = auth()->guard('researcher')->user();
-
-        if ($researcher && $researcher->currentAccessToken()) {
-            $researcher->currentAccessToken()->delete();
+        $company = auth()->guard('company')->user();
+    
+        if ($company && $company->currentAccessToken())
+        {
+            $company->currentAccessToken()->delete();
+            return response()->json(['message' => 'تم تسجيل الخروج بنجاح']);    
         }
-        return response()->json(['message' => 'تم تسجيل الخروج بنجاح']);
-    }
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(Request $request)
-    {
-        //
+
     }
 
-    /**
-     * Display the specified resource.
-     */
-    public function show(string $id)
-    {
-        //
-    }
 
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, string $id)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(string $id)
-    {
-        //
-    }
+   
 }
+
