+++ app/Http/Controllers/Api/Report/ReportController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Api\Report;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\ProductResource;
